Ignore volumes when constructing Koji URLs
Guessing the volume name is not reliable and things can break
unexpectedly in future.
This happened recently in Fedora [1].

Luckily for us, Koji maintains symlinks on the default volume
for builds on other volumes [2]. This means that we should be able
to just ignore the volume in the URL completely.

For example:
${BREWWEB_ROOT}/vol/rhel-9/packages/systemd/250/10.el9/x86_64/systemd-250-10.el9.x86_64.rpm
We can use just:
${BREWWEB_ROOT}/packages/systemd/250/10.el9/x86_64/systemd-250-10.el9.x86_64.rpm

[1]: https://pagure.io/fedora-ci/general/issue/356
[2]: https://github.com/koji-project/koji/blob/master/docs/source/volumes.rst#symlinks-on-default-volume

Signed-off-by: Michal Srb <michal@redhat.com>