Add "--scriptletcheck" option
YUM operations can succeed, but there can be scriptlets errors
in the output.

This commit adds a new "--scriptletcheck" option that tells mini-tps
to check the output of the YUM command and fail the test if any
problematic words are found.

It's hard to tell how effective this approach will be in the real
world -- only thorough testing will tell.

Signed-off-by: Michal Srb <michal@redhat.com>