Refactor Brew querying functions
They are now brew_ prefixed and they not only prepare the query
but also send it.

Most of the *_from_answer() functions have been renamed to <something>2<attribute>()
where <something> is what (XML) we got from Brew
and <attribute> is what we want to extract from it.
The 'from_answer' in the function names had no information value.