Differentiate between 'no x86_64/noarch pkgs' and 'no pkgs available in repos'
The 'no x86_64/noarch pkgs' means the build contains only packages of an alternative arch.

The 'no pkgs available in repos' means that there are x86_64/noarch
packages but none of them is in a repo so it's probably a new package.

Related to https://issues.redhat.com/browse/OSCI-5765