fall back on `{dnf,yum} repoquery` if `repoquery` is not present (#47)
https://github.com/fedora-ci/mini-tps/pull/45 changed two out of
three places we previously called `"$YUMDNFCMD" repoquery` to
instead just call `repoquery` (it missed one). This is because
`yum repoquery` doesn't exist on EL 7. However, it's also not
really safe to assume the standalone `repoquery` command will
always be available either; it's part of yum-utils or dnf-utils
and there's no reason to assume that will always be present. This
adds a fallback, and consistently uses the command we determine
across all three locations.

Signed-off-by: Adam Williamson <awilliam@redhat.com>