consistently call plain `repoquery`, ensure it is available (#47)
https://github.com/fedora-ci/mini-tps/pull/45 changed two of the
three places we try and do a repoquery operation from calling
`"$YUMDNFCMD" repoquery` to just calling `repoquery`, but it
missed one. This fixes the remaining one. Also, this attempts
to ensure the plain `repoquery` command will always be available.
It's not safe to assume this, because it's part of dnf-utils (on
Fedora) and yum-utils (on EL), it's not part of core dnf/yum, and
nothing much else requires it. It does seem to be missing in the
environment in which tests on Fedora package pull requests run,
for instance, which I think causes the test to fail on all pull
requests at present.

Signed-off-by: Adam Williamson <awilliam@redhat.com>