Determine if an imported build is a draft
Currently, we check the task to see if the build is a draft. This
does not work for import tasks, as those tasks do not necessarily
contain this information (and certainly don't for Konflux).
Instead, once we have determined that the task is an import,
consult the build to determine whether it is a draft or not and
then proceed accordingly.

Resolves: OSCI-8507